--- app/models/foods/index.ts
@@ -1,1 +1,0 @@
-export * from "./Food.ts";
