+++ app/models/index.ts
@@ -1,2 +1,2 @@
-export * from "./foods/index.ts";
+export * from "./recipes/index.ts";
 export * from "./nutrition/index.ts";
